[frontend] StixCyberObservableDetails to tsx
